renaming schools list to browse
