+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/BrowseActivity.java
@@ -1,23 +1,24 @@
-package blueprint.com.sage.schools;
+package blueprint.com.sage.browse;
 
 import android.os.Bundle;
 
 import com.google.android.gms.common.api.GoogleApiClient;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 import blueprint.com.sage.R;
 import blueprint.com.sage.models.School;
 import blueprint.com.sage.network.Requests;
-import blueprint.com.sage.schools.fragments.SchoolListFragment;
+import blueprint.com.sage.browse.fragments.SchoolListFragment;
 import blueprint.com.sage.shared.activities.NavigationAbstractActivity;
 import blueprint.com.sage.utility.view.FragUtils;
 
 /**
  * Created by charlesx on 11/4/15.
  */
-public class SchoolsListActivity extends NavigationAbstractActivity {
+public class BrowseActivity extends NavigationAbstractActivity {
 
     private List<School> mSchools;
     private GoogleApiClient mGoogleApiClient;
@@ -29,11 +30,18 @@ public class SchoolsListActivity extends NavigationAbstractActivity {
         mSchools = new ArrayList<>();
 
         getSchoolsListRequest();
+        getUsersListRequest();
         FragUtils.replace(R.id.container, SchoolListFragment.newInstance(), this);
     }
 
     public void getSchoolsListRequest() {
         Requests.Schools.with(this).makeListRequest(null);
+    }
+    public void getUsersListRequest() {
+        HashMap<String, String> queryParams = new HashMap<>();
+        queryParams.put("verified", "true");
+
+        Requests.Users.with(this).makeListRequest(queryParams);
     }
 
     public void setSchools(List<School> schools) { mSchools = schools; }
