+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/adapters/SchoolsListAdapter.java
@@ -1,4 +1,4 @@
-package blueprint.com.sage.schools.adapters;
+package blueprint.com.sage.browse.adapters;
 
 import android.app.Activity;
 import android.support.v7.widget.RecyclerView;
