+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/CreateSchoolFragment.java
@@ -1,4 +1,4 @@
-package blueprint.com.sage.schools.fragments;
+package blueprint.com.sage.browse.fragments;
 
 import android.os.Bundle;
 import android.text.Editable;
@@ -27,7 +27,7 @@ import blueprint.com.sage.R;
 import blueprint.com.sage.events.schools.CreateSchoolEvent;
 import blueprint.com.sage.models.School;
 import blueprint.com.sage.network.Requests;
-import blueprint.com.sage.schools.adapters.PlacePredictionAdapter;
+import blueprint.com.sage.browse.adapters.PlacePredictionAdapter;
 import blueprint.com.sage.shared.FormValidation;
 import blueprint.com.sage.utility.view.FragUtils;
 import blueprint.com.sage.utility.view.MapUtils;
