+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/SchoolAbstractFragment.java
@@ -1,0 +1,12 @@
+package blueprint.com.sage.browse.fragments;
+
+import android.support.v4.app.Fragment;
+
+import blueprint.com.sage.browse.BrowseActivity;
+
+/**
+ * Created by charlesx on 11/4/15.
+ */
+public class SchoolAbstractFragment extends Fragment {
+    public BrowseActivity getParentActivity() { return (BrowseActivity) getActivity(); }
+}
