+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/SchoolListFragment.java
@@ -1,4 +1,4 @@
-package blueprint.com.sage.schools.fragments;
+package blueprint.com.sage.browse.fragments;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -11,7 +11,7 @@ import android.view.ViewGroup;
 
 import blueprint.com.sage.R;
 import blueprint.com.sage.events.schools.SchoolListEvent;
-import blueprint.com.sage.schools.adapters.SchoolsListAdapter;
+import blueprint.com.sage.browse.adapters.SchoolsListAdapter;
 import blueprint.com.sage.shared.views.RecycleViewEmpty;
 import blueprint.com.sage.utility.view.FragUtils;
 import butterknife.Bind;
