--- sage-android/Sage/app/src/main/java/blueprint/com/sage/schools/fragments/SchoolAbstractFragment.java
@@ -1,12 +1,0 @@
-package blueprint.com.sage.schools.fragments;
-
-import android.support.v4.app.Fragment;
-
-import blueprint.com.sage.schools.SchoolsListActivity;
-
-/**
- * Created by charlesx on 11/4/15.
- */
-public class SchoolAbstractFragment extends Fragment {
-    public SchoolsListActivity getParentActivity() { return (SchoolsListActivity) getActivity(); }
-}
